Secure HTTP requests. Validate context HTTP retrieval.
- Handle HTTP errors when retrieving contexts by URL on "upload
specification" phase.
