+++ app/controllers/concerns/connectable.rb
@@ -29,6 +29,9 @@ module Connectable
     ###
     def http_get uri
       HTTParty.get(uri)
+    rescue StandardError => e
+      puts e.inspect
+      false
     end
   end
 end
